+++ src/materials.rs
@@ -19,8 +19,8 @@ pub enum MaterialType {
 #[derive(Copy, Clone)]
 pub struct Material {//Used in:
     pub albedo: Color,//Lambertian, Metal
-    pub fuzz: f64,//Metal
-    pub ior: f64,//Dielectric
+    pub fuzz: f32,//Metal
+    pub ior: f32,//Dielectric
     pub mat_type: MaterialType, //Tag
 }
 
@@ -31,10 +31,10 @@ impl Material{
     pub fn new_metal(albedo: Color) -> Self{
         return Self{albedo: albedo,fuzz: 0.,ior: 0.,mat_type: MaterialType::METAL};
     }
-    pub fn new_metal_fuzz(albedo: Color,fuzz: f64) -> Self{
+    pub fn new_metal_fuzz(albedo: Color,fuzz: f32) -> Self{
         return Self{albedo: albedo,fuzz: fuzz,ior: 0.,mat_type: MaterialType::METAL};
     }
-    pub fn new_dielectric(index_of_refraction: f64) -> Self{
+    pub fn new_dielectric(index_of_refraction: f32) -> Self{
         return Self{albedo: Color::ZERO,fuzz: 0.,ior: index_of_refraction,mat_type: MaterialType::DIELECTRIC};
     }
     pub fn scatter(&self,r_in: &Ray,hr: &HitRecord) -> MaterialScatterResult {
@@ -69,7 +69,7 @@ impl Material{
     pub fn scatter_dielectric(&self,r_in: &Ray,hr: &HitRecord) -> MaterialScatterResult {
         let dir_unit = r_in.dir.unit();
         let front_face = dir_unit.dot(hr.normal) < 0.;
-        let refraction_ratio: f64;
+        let refraction_ratio: f32;
         let normal: Vec3;
         if front_face {
             refraction_ratio = 1.0/self.ior;
@@ -83,7 +83,7 @@ impl Material{
         let cos_theta = min((-dir_unit).dot(normal),1.0);
         let sin_theta = (1.0 - cos_theta*cos_theta).sqrt();
         let cannot_refract = (refraction_ratio * sin_theta) > 1.0;
-        let reflect_by_reflectante = reflectance(cos_theta, refraction_ratio) > f64::rand();
+        let reflect_by_reflectante = reflectance(cos_theta, refraction_ratio) > f32::rand();
         let new_dir: Vec3;
         if cannot_refract || reflect_by_reflectante {//Reflect
             new_dir = reflect(&dir_unit, &normal);
@@ -100,7 +100,7 @@ fn reflect(v: &Vec3,n: &Vec3) -> Vec3{
     return *v - 2.*v.dot(*n)*(*n);
 }
 
-fn refract(uv: &Vec3,n: &Vec3,etai_over_etat: f64) -> Vec3 {
+fn refract(uv: &Vec3,n: &Vec3,etai_over_etat: f32) -> Vec3 {
     let cos_theta = min((-*uv).dot(*n),1.0);
     let r_out_perp = etai_over_etat * ((*uv) + cos_theta*(*n));
     let aux = -(1.0 - r_out_perp.length_squared()).abs().sqrt();
@@ -108,7 +108,7 @@ fn refract(uv: &Vec3,n: &Vec3,etai_over_etat: f64) -> Vec3 {
     return r_out_perp + r_out_parallel;
 }
 
-fn reflectance(cos: f64, ref_idx: f64) -> f64{
+fn reflectance(cos: f32, ref_idx: f32) -> f32{
     // Use Schlick's approximation for reflectance.
     let r0 = (1.-ref_idx)/(1.+ref_idx);
     let r0_2 = r0*r0;
